chore: Update .NET SDK version to 9.0.105
This commit updates the .NET SDK version used in the project from 9.0.104 to 9.0.105. This change is reflected in the `runtimeIdentifierGraphPath` property within the `*.csproj.nuget.dgspec.json` files for PocTests, PocApplication and PocInfra. Additionally, AssemblyInformationalVersion is updated in PocInfra, PocDomain and PocApplication. The project.nuget.cache and project.assets.json files are also updated to reflect this change.